+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [0.7.0] - 2026-##-##
+
+### Added
+- Shared service layer for customers and invoices
+- Django REST Framework API layer
+- Customer and invoice API endpoints
+- API tests for customer and invoice endpoints
+
+### Changed
+- Modify cli tests to reflect service-layer archtiecture 
+
 ## [0.6.0] - 2026-04-14
 
 ### Added